requirement.txt
streamlit
langchain
python-dotenv
huggingface_hub
openai